+++ src/wurf/git_semver_resolver.py
@@ -12,22 +12,23 @@ class GitSemverResolver(object):
     Read more about Semantic Versioning here: semver.org
     """
 
-    def __init__(self, git, git_resolver, ctx, semver, bundle_path, dependency):
+    def __init__(self, git, git_resolver, ctx, semver_selector, working_path,
+        dependency):
         """ Construct an instance.
 
         :param git: A WurfGit instance
         :param url_resolver: A WurfGitResolver instance.
         :param ctx: A Waf Context instance.
-        :param semver: The semver module
-        :param bundle_path: Current working directory as a string. This is the place
+        :param semver_selector: A SemverSelector instance.
+        :param working_path: Current working directory as a string. This is the place
             where we should create new folders etc.
         :param dependency: The dependency instance.
         """
         self.git = git
         self.git_resolver = git_resolver
         self.ctx = ctx
-        self.semver = semver
-        self.bundle_path = bundle_path
+        self.semver_selector = semver_selector
+        self.working_path = working_path
         self.dependency = dependency
 
     def resolve(self):
@@ -40,76 +41,38 @@ class GitSemverResolver(object):
 
         assert os.path.isdir(path)
 
+        tags = self.git.tags(cwd=path)
+        tag = self.semver_selector.select_tag(
+            major=self.dependency.major, tags=tags)
+
+        if not tag:
+            self.ctx.fatal(
+                'No major tag {} for {} found. Candiates were: {}'.format(
+                    self.dependency.major, self.dependency.name, tags))
+
         # Use the path retuned to create a unique location for this checkout
         repo_hash = hashlib.sha1(path.encode('utf-8')).hexdigest()[:6]
 
-        tags = self.git.tags(cwd=path)
-        tag = self.select_tag(major=self.dependency.major, tags=tags)
-
-        if not tag:
-            self.ctx.fatal('No major tag {} for {} found. Candiates were: {}'.format(
-            self.dependency.major, self.dependency.name, tags))
-
-        # Record the selected tag
-        self.dependency.git_tag = tag
-
         # The folder for storing different versions of this repository
-        repo_name = self.dependency.name + '-' + tag + '-' + repo_hash
-        repo_path = os.path.join(self.bundle_path, repo_name)
+        repo_name = tag + '-' + repo_hash
+        repo_path = os.path.join(self.working_path, repo_name)
 
         self.ctx.to_log('wurf: GitSemverResolver name {} -> {}'.format(
             self.dependency.name, repo_path))
 
-        # If the checkout folder does not exist,
-        # then clone from the git repository
+        # If the folder for the chosen tag does not exist,
+        # then copy the master and checkout the tag
         if not os.path.isdir(repo_path):
             shutil.copytree(src=path, dst=repo_path, symlinks=True)
             self.git.checkout(branch=tag, cwd=repo_path)
 
-        # If the project contains submodules we also get those
-        #
+        # If the project contains submodules, we also get those
         self.git.pull_submodules(cwd=repo_path)
 
+        # Record the commmit id of the current working copy
+        self.dependency.git_commit = self.git.current_commit(cwd=repo_path)
+
         return repo_path
-
-    def select_tag(self, major, tags):
-        """
-        Compare the available tags and return the newest tag for the
-        specific version.
-
-        :param tags: list of available tags
-        :return: the tag to use or None if not tag is compatible
-        """
-        assert isinstance(major, int), "Major version is not an int"
-
-        # Get tags with matching version
-        valid_tags = []
-
-        for tag in tags:
-            try:
-                t = self.semver.parse(tag)
-                if t['major'] != major:
-                    continue
-
-                valid_tags.append(tag)
-            except ValueError:  # ignore tags we cannot parse
-                pass
-
-        if len(valid_tags) == 0:
-            return None
-
-        # Now figure out which version is the newest. We may only
-        # use versions meeting the version requirement as specified by
-        # self.major, self.minor, and self.patch, to ensure compatibility see
-        # rules at semver.org
-
-        best_match = valid_tags[0]
-
-        for t in valid_tags:
-            if self.semver.match(best_match, "<" + t):
-                best_match = t
-
-        return best_match
 
     def __repr__(self):
         """