+++ src/wurf/on_passive_load_path_resolver.py
@@ -8,15 +8,15 @@ from .error import DependencyError
 
 class OnPassiveLoadPathResolver(object):
 
-    def __init__(self, dependency, bundle_config_path):
+    def __init__(self, dependency, resolve_config_path):
         """ Construct an instance.
 
         :param dependency: A Dependency instance.
-        :param bundle_config_path: A string containing the path to where the
+        :param resolve_config_path: A string containing the path to where the
             dependencies config json files should be / is stored.
         """
         self.dependency = dependency
-        self.bundle_config_path = bundle_config_path
+        self.resolve_config_path = resolve_config_path
 
     def resolve(self):
         """ Resolve a path to a dependency.
@@ -50,7 +50,7 @@ class OnPassiveLoadPathResolver(object):
         """
 
         config_path = os.path.join(
-            self.bundle_config_path, self.dependency.name + '.resolve.json')
+            self.resolve_config_path, self.dependency.name + '.resolve.json')
 
         if not os.path.isfile(config_path):
             raise DependencyError('No config - re-run configure',