+++ src/wurf/options.py
@@ -5,7 +5,7 @@ from .error import Error
 
 class Options(object):
 
-    def __init__(self, args, parser, default_bundle_path,
+    def __init__(self, args, parser, default_resolve_path,
                  default_symlinks_path, supported_git_protocols):
 
         self.args = args
@@ -16,24 +16,24 @@ class Options(object):
 
         # Using the %default placeholder:
         #    http://stackoverflow.com/a/1254491/1717320
-        self.parser.add_argument('--bundle-path',
-            default=default_bundle_path,
-            dest='--bundle-path',
+        self.parser.add_argument('--resolve_path',
+            default=default_resolve_path,
+            dest='--resolve_path',
             help='The folder where the bundled dependencies are downloaded.'
                  '(default: %(default)s)')
 
-        self.parser.add_argument('--git-protocol',
-            dest='--git-protocol',
+        self.parser.add_argument('--git_protocol',
+            dest='--git_protocol',
             help='Use a specific git protocol to download dependencies.'
                  'Supported protocols {}'.format(supported_git_protocols))
 
-        self.parser.add_argument('--symlinks-path',
+        self.parser.add_argument('--symlinks_path',
             default=default_symlinks_path,
-            dest='--symlinks-path',
+            dest='--symlinks_path',
             help='The folder where the dependency symlinks are placed.'
                  '(default: %(default)s)')
 
-        self.parser.add_argument('--fast-resolve', dest='--fast-resolve',
+        self.parser.add_argument('--fast_resolve', dest='--fast_resolve',
             action='store_true', default=False,
             help='Load already resolved dependencies from file system. '
                   'Useful for running configure without resolving dependencies '
@@ -51,17 +51,17 @@ class Options(object):
 
         self.__parse()
 
-    def bundle_path(self):
-        return self.known_args['--bundle-path']
+    def resolve_path(self):
+        return self.known_args['--resolve_path']
 
     def symlinks_path(self):
-        return self.known_args['--symlinks-path']
+        return self.known_args['--symlinks_path']
 
     def git_protocol(self):
-        return self.known_args['--git-protocol']
+        return self.known_args['--git_protocol']
 
     def fast_resolve(self):
-        return self.known_args['--fast-resolve']
+        return self.known_args['--fast_resolve']
 
     def lock_paths(self):
         return self.known_args['--lock_paths']
@@ -70,10 +70,10 @@ class Options(object):
         return self.known_args['--lock_versions']
 
     def path(self, dependency):
-        return self.known_args['--%s-path' % dependency.name]
+        return self.known_args['--%s_path' % dependency.name]
 
-    def use_checkout(self, dependency):
-        return self.known_args['--%s-use-checkout' % dependency.name]
+    def checkout(self, dependency):
+        return self.known_args['--%s_checkout' % dependency.name]
 
     def __parse(self):
 
@@ -87,7 +87,7 @@ class Options(object):
 
     def __add_path(self, dependency):
 
-        option = '--%s-path' % dependency.name
+        option = '--%s_path' % dependency.name
 
         self.parser.add_argument(option,
             nargs='?',
@@ -95,9 +95,9 @@ class Options(object):
             help='Manually specify path for {}.'.format(
                 dependency.name))
 
-    def __add_use_checkout(self, dependency):
+    def __add_checkout(self, dependency):
 
-        option = '--%s-use-checkout' % dependency.name
+        option = '--%s_checkout' % dependency.name
 
         self.parser.add_argument(option,
             nargs='?',
@@ -111,6 +111,6 @@ class Options(object):
 
         if dependency.resolver == 'git':
 
-            self.__add_use_checkout(dependency)
+            self.__add_checkout(dependency)
 
         self.__parse()