+++ src/wurf/waf_resolve_context.py
@@ -36,15 +36,13 @@ figure out which dependencies exist.
 """
 
 class WafResolveContext(Context.Context):
-
     '''resolves the dependencies specified in the wscript's resolve function'''
 
     cmd = 'resolve'
     fun = 'resolve'
 
     def __init__(self, **kw):
-        """ Create a WafResolveContext
-        """
+        """ Create a WafResolveContext"""
         super(WafResolveContext, self).__init__(**kw)
 
     def execute(self):
@@ -63,16 +61,16 @@ class WafResolveContext(Context.Context):
         self.bldnode = self.path.make_node('build')
         self.bldnode.mkdir()
 
-        default_bundle_path = os.path.join(
-            self.path.abspath(), 'bundle_dependencies')
+        default_resolve_path = os.path.join(
+            self.path.abspath(), 'resolved_dependencies')
 
         default_symlinks_path = os.path.join(
             self.path.abspath(), 'build_symlinks')
 
         self.registry = registry.build_registry(
             ctx=self, git_binary='git',
-            semver=semver, default_bundle_path=default_bundle_path,
-            bundle_config_path=self.bundle_config_path(),
+            semver=semver, default_resolve_path=default_resolve_path,
+            resolve_config_path=self.resolve_config_path(),
             default_symlinks_path=default_symlinks_path,
             waf_utils=Utils, args=sys.argv[1:],
             project_path=self.path.abspath())
@@ -100,13 +98,13 @@ class WafResolveContext(Context.Context):
             # which will trigger loading the dependency.
             super(WafResolveContext, self).execute()
 
-            # Now try to load dependencies from the resolve.json file. Note,
-            # that this is done after calling a possible user-defined
-            # resolve(...) function. Since we always want the allow the user to
-            # run custom code before we start resolving stuff.
+            # As the last step in recurse, try to load the dependencies from the
+            # 'resolve.json' file if it is present next to the wscript.
+            # This is done after calling a possible user-defined
+            # resolve() function, since we always want to allow the user to
+            # run custom code before the actual resolving starts.
             self.dependency_manager.load_dependencies(self.path.abspath(),
                 mandatory=False)
-
         except Error as e:
             self.logger.debug("Error in resolve:".format(e), exc_info=True)
             self.fatal("Error: {}".format(e))
@@ -132,25 +130,19 @@ class WafResolveContext(Context.Context):
         """
         return self.srcnode == self.path
 
-
-    def bundle_config_path(self):
+    def resolve_config_path(self):
         """Returns the bundle config path.
 
         The bundle config path will be used to store/load configuration for
         the different dependencies that are resolved.
         """
-
         return self.bldnode.abspath()
 
-
     def add_dependency(self, **kwargs):
-        """Adds a dependency.
-        """
-
+        """Adds a dependency."""
         self.dependency_manager.add_dependency(**kwargs)
 
     def cmd_and_log(self, cmd, **kwargs):
-
         # Seems the new waf needs the cwd to be a Node object. We do this
         # adaption here to avoid introducing additional Waf dependencies in the
         # other parts of the code.