+++ test/python/test_dependency.py
@@ -3,15 +3,18 @@ import mock
 
 from wurf.dependency import Dependency
 
-def test_wurf_dependency():
+def test_dependency():
 
-    dep = {"name":"waf",
-           "optional": True,
-           "recurse": False,
-           "resolver": "git",
-           "method": "checkout",
-           "checkout": "1.3.3.7",
-           "sources":["gitrepo1.git", "gitrepo2.git"]}
+    dep = \
+    {
+        "name": "waf",
+        "optional": True,
+        "recurse": False,
+        "resolver": "git",
+        "method": "checkout",
+        "checkout": "somebranch",
+        "sources": ["gitrepo1.git", "gitrepo2.git"]
+    }
 
     w = Dependency(**dep)
 
@@ -20,13 +23,14 @@ def test_wurf_dependency():
     assert w.recurse == False
     assert w.resolver == "git"
     assert w.method == "checkout"
+    assert w.checkout == "somebranch"
     assert w.sources == ["gitrepo1.git", "gitrepo2.git"]
 
     # Check there is a sha1 key
     assert "sha1" in w
     assert len(w.sha1) > 0
 
-    # Check that we change the read-only attributes (the ones passed
+    # Check that we cannot change the read-only attributes (the ones passed
     # at construction time)
     with pytest.raises(Exception):
         w.name = "waf2"
@@ -40,3 +44,35 @@ def test_wurf_dependency():
     w.rewrite(attribute='method', value='semver', reason='testing it')
     w.rewrite(attribute='major', value=3, reason='testing it')
     w.rewrite(attribute='checkout', value=None, reason='testing it')
+
+    assert w.method == "semver"
+    assert w.major == 3
+    assert 'checkout' not in w
+
+
+def test_dependency_default_values():
+
+    dep = \
+    {
+        "name": "foo",
+        "resolver": "git",
+        "method": "checkout",
+        "sources": ["gitrepo1.git", "gitrepo2.git"]
+    }
+
+    w = Dependency(**dep)
+
+    # Check the explicitly defined attributes
+    assert w.name == "foo"
+    assert w.resolver == "git"
+    assert w.method == "checkout"
+    assert w.sources == ["gitrepo1.git", "gitrepo2.git"]
+
+    # Check the default values
+    assert w.recurse == True
+    assert w.optional == False
+    assert w.internal == False
+
+    # Check there is a sha1 key
+    assert "sha1" in w
+    assert len(w.sha1) > 0