+++ test/python/test_git_resolver.py
@@ -9,12 +9,12 @@ def test_git_resolver(test_directory):
 
     ctx = mock.Mock()
     git = mock.Mock()
-    cwd = test_directory.path()
+    working_path = test_directory.path()
 
     # GitResolver checks that the directory is created during git.clone,
     # so we create it within the test_directory as a side effect
     def fake_git_clone(repository, directory, cwd):
-        test_directory.mkdir(directory)
+        os.makedirs(os.path.join(cwd, directory))
 
     git.clone = mock.Mock(side_effect=fake_git_clone)
 
@@ -22,16 +22,16 @@ def test_git_resolver(test_directory):
     repo_url = 'https://gitlab.com/steinwurf/links.git'
 
     resolver = GitResolver(git=git, ctx=ctx, name=name,
-        bundle_path=cwd, source=repo_url)
+        working_path=working_path, source=repo_url)
 
     path = resolver.resolve()
 
-    # The destination path is determined in the resolve function,
-    # we just get the folder name manually
     repo_name = os.path.basename(os.path.normpath(path))
+    assert repo_name.startswith('master-')
+    repo_folder = os.path.dirname(os.path.normpath(path))
 
     git.clone.assert_called_once_with(
-        repository=repo_url, directory=repo_name, cwd=cwd)
+        repository=repo_url, directory=repo_name, cwd=repo_folder)
 
     git.pull_submodules.assert_called_once_with(cwd=path)