+++ test/python/test_options.py
@@ -5,27 +5,27 @@ import argparse
 from wurf.registry import Options
 
 
-def test_bundle_path():
+def test_resolve_path():
 
-    default_path = 'bundle_dependencies'
+    default_path = 'resolved_dependencies'
 
     parser = argparse.ArgumentParser()
     args = ['--foo', '-b']
 
     options = Options(args=args, parser=parser,
-        default_bundle_path=default_path,
+        default_resolve_path=default_path,
         default_symlinks_path="", supported_git_protocols="")
 
-    assert options.bundle_path() == default_path
+    assert options.resolve_path() == default_path
 
 
     parser = argparse.ArgumentParser()
-    args = ['--foo', '--bundle-path', '/tmp/bundlehere', '-b']
+    args = ['--foo', '--resolve_path', '/tmp/bundlehere', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path="", supported_git_protocols="")
 
-    assert options.bundle_path() == '/tmp/bundlehere'
+    assert options.resolve_path() == '/tmp/bundlehere'
 
     dependency = mock.Mock()
     dependency.name = 'foo'
@@ -42,15 +42,15 @@ def test_symlinks_path():
     parser = argparse.ArgumentParser()
     args = ['--foo', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path=default_path, supported_git_protocols="")
 
     assert options.symlinks_path() == default_path
 
     parser = argparse.ArgumentParser()
-    args = ['--foo', '--symlinks-path', '/tmp/symlinks', '-b']
+    args = ['--foo', '--symlinks_path', '/tmp/symlinks', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path=default_path, supported_git_protocols="")
 
     assert options.symlinks_path() == '/tmp/symlinks'
@@ -63,10 +63,10 @@ def test_user_path_to_dependency():
 
     # Path specified
     parser = argparse.ArgumentParser()
-    args = ['--foo', '--bundle-path', '/tmp/bundlehere',
-            '--foo-path', '/home/stw/code', '-b']
+    args = ['--foo', '--resolve_path', '/tmp/bundlehere',
+            '--foo_path', '/home/stw/code', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path="", supported_git_protocols="")
 
     options.add_dependency(dependency)
@@ -75,10 +75,10 @@ def test_user_path_to_dependency():
 
     # Path specified
     parser = argparse.ArgumentParser()
-    args = ['--foo', '--bundle-path', '/tmp/bundlehere',
-            '--foo-path', '/home/stw/code1', '-b']
+    args = ['--foo', '--resolve_path', '/tmp/bundlehere',
+            '--foo_path', '/home/stw/code1', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path="", supported_git_protocols="")
 
     options.add_dependency(dependency)
@@ -91,15 +91,15 @@ def test_git_protocol():
     parser = argparse.ArgumentParser()
     args = ['--foo', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path="", supported_git_protocols="")
 
     assert options.git_protocol() == None
 
     parser = argparse.ArgumentParser()
-    args = ['--foo', '--git-protocol', 'myproto://', '-b']
+    args = ['--foo', '--git_protocol', 'myproto://', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path="", supported_git_protocols="")
 
     assert options.git_protocol() == 'myproto://'
@@ -110,15 +110,15 @@ def test_fast_resolve():
     parser = argparse.ArgumentParser()
     args = ['--foo', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path="", supported_git_protocols="")
 
     assert options.fast_resolve() == False
 
     parser = argparse.ArgumentParser()
-    args = ['--foo', '--fast-resolve', '-b']
+    args = ['--foo', '--fast_resolve', '-b']
 
-    options = Options(args=args, parser=parser, default_bundle_path="",
+    options = Options(args=args, parser=parser, default_resolve_path="",
         default_symlinks_path="", supported_git_protocols="")
 
     assert options.fast_resolve() == True